Study/telegram_backup/requirements.txt
cryptg==0.4.0
markdown-it-py==3.0.0
mdurl==0.1.2
pyaes==1.6.1
pyasn1==0.5.1
Pygments==2.17.2
PyYAML==6.0.1
rich==13.7.0
rsa==4.9
Telethon==1.34.0
tqdm==4.66.1
